refactor : env 개발 / 배포 파일 분리, Share 버튼 컴포넌트 분리
