+++ .env.development
@@ -1,0 +1,6 @@
+NEXT_PUBLIC_KAKAO_RESTAPI_KEY = 689817d74a38736df40afee3ec9578d8
+NEXT_PUBLIC_KAKAO_JAVASCRIPT_KEY = 0e0d4582f91b667c231b66a8297d4ddb
+
+NEXT_PUBLIC_FRONTEND = http://localhost:3000
+NEXT_PUBLIC_BACKEND = https://dev.catchcatch.link/v1
+NEXT_PUBLIC_KAKAO_REDIRECT_URI = http://localhost:3000/redirect
