+++ components/share/index.ts
@@ -1,0 +1,6 @@
+export { default as ShareModalBtn } from './ShareModalBtn';
+export { default as Facebook } from './Facebook';
+export { default as Twitter } from './Twitter';
+export { default as Kakaotalk } from './Kakaotalk';
+export { default as ShareModal } from './ShareModal';
+export { default as LinkCopyBtn } from './LinkCopyBtn';
