+++ pages/quiz/solve/[quizset_id]/result/[solver_id]/index.tsx
@@ -9,8 +9,8 @@ import { theme } from 'styles/theme';
 import UserScore from 'components/resultPage/UserScore';
 import { Emotion } from 'components/emotion';
 import Sketchbook from 'components/style/Sketchbook';
-import ShareBtn from 'components/common/ShareBtn';
 import Comment from 'components/comment/Comment';
+import { ShareModalBtn } from 'components/share';
 
 const Page: NextPageWithLayout = () => {
   const { solveUserName, solveUserScore } = useSelector((state: RootState) => state.user_solve);
@@ -24,7 +24,7 @@ const Page: NextPageWithLayout = () => {
             <UserScore name={solveUserName} score={solveUserScore} total={quizList.length} />
             <RankingBoard />
             <Emotion />
-            <ShareBtn />
+            <ShareModalBtn />
             <Comment />
           </Wrapper>
         </Sketchbook>
